Resources: refactor etag handling inside utils
