+++ azure-mgmt-dns/src/main/java/com/azure/management/dns/implementation/DnsRecordSetImpl.java
@@ -19,6 +19,7 @@ import com.azure.management.dns.PtrRecord;
 import com.azure.management.dns.RecordType;
 import com.azure.management.dns.SrvRecord;
 import com.azure.management.dns.TxtRecord;
+import com.azure.management.resources.fluentcore.utils.ETagState;
 import reactor.core.publisher.Mono;
 
 import java.util.ArrayList;
@@ -408,63 +409,5 @@ class DnsRecordSetImpl extends ExternalChildResourceImpl<DnsRecordSet,
     DnsRecordSetImpl withETagOnDelete(String eTagValue) {
         this.eTagState.withExplicitETagCheckOnDelete(eTagValue);
         return this;
-    }
-
-    private class ETagState {
-        private boolean doImplicitETagCheckOnCreate;
-        private boolean doImplicitETagCheckOnUpdate;
-        private String eTagOnUpdate;
-        private String eTagOnDelete;
-
-        public ETagState withImplicitETagCheckOnCreate() {
-            this.doImplicitETagCheckOnCreate = true;
-            return this;
-        }
-
-        public ETagState withImplicitETagCheckOnUpdate() {
-            this.doImplicitETagCheckOnUpdate = true;
-            return this;
-        }
-
-        public ETagState withExplicitETagCheckOnUpdate(String eTagValue) {
-            this.eTagOnUpdate = eTagValue;
-            return this;
-        }
-
-        public ETagState withExplicitETagCheckOnDelete(String eTagValue) {
-            this.eTagOnDelete = eTagValue;
-            return this;
-        }
-
-
-        public ETagState clear() {
-            this.doImplicitETagCheckOnCreate = false;
-            this.doImplicitETagCheckOnUpdate = false;
-            this.eTagOnUpdate = null;
-            this.eTagOnDelete = null;
-            return this;
-        }
-
-        public String ifMatchValueOnUpdate(String currentETagValue) {
-            String eTagValue = null;
-            if (this.doImplicitETagCheckOnUpdate) {
-                eTagValue = currentETagValue;
-            }
-            if (this.eTagOnUpdate != null) {
-                eTagValue = this.eTagOnUpdate;
-            }
-            return eTagValue;
-        }
-
-        public String ifMatchValueOnDelete() {
-            return this.eTagOnDelete;
-        }
-
-        public String ifNonMatchValueOnCreate() {
-            if (this.doImplicitETagCheckOnCreate) {
-                return "*";
-            }
-            return null;
-        }
     }
 }
