+++ azure-mgmt-dns/src/main/java/com/azure/management/dns/implementation/DnsZoneImpl.java
@@ -11,6 +11,7 @@ import com.azure.core.management.SubResource;
 import com.azure.management.dns.models.RecordSetInner;
 import com.azure.management.dns.models.ZoneInner;
 import com.azure.management.resources.fluentcore.arm.models.implementation.GroupableResourceImpl;
+import com.azure.management.resources.fluentcore.utils.ETagState;
 import com.azure.management.resources.fluentcore.utils.PagedConverter;
 import com.azure.management.resources.fluentcore.utils.Utils;
 import com.azure.management.dns.ARecordSets;
@@ -56,7 +57,7 @@ public class DnsZoneImpl
     private SrvRecordSets srvRecordSets;
     private TxtRecordSets txtRecordSets;
     private DnsRecordSetsImpl recordSets;
-    private String dnsZoneETag;
+    private final ETagState eTagState = new ETagState();
 
     DnsZoneImpl(String name, final ZoneInner innerModel, final DnsZoneManager manager) {
         super(name, innerModel, manager);
@@ -397,34 +398,25 @@ public class DnsZoneImpl
 
     @Override
     public DnsZoneImpl withETagCheck() {
-        if (isInCreateMode()) {
-            this.dnsZoneETag = "*";
-            return this;
-        }
-        return this.withETagCheck(this.inner().etag());
+        this.eTagState.withImplicitETagCheckOnCreate();
+        this.eTagState.withImplicitETagCheckOnUpdate();
+        return this;
     }
 
     @Override
     public DnsZoneImpl withETagCheck(String eTagValue) {
-        this.dnsZoneETag = eTagValue;
+        this.eTagState.withExplicitETagCheckOnUpdate(eTagValue);
         return this;
     }
 
     @Override
     public Mono<DnsZone> createResourceAsync() {
         return Mono.just(this)
-                .flatMap(self -> {
-                    if (self.isInCreateMode()) {
-                        return self.manager().inner().zones().createOrUpdateAsync(self.resourceGroupName(),
-                                self.name(), self.inner(), null/*IfMatch*/, self.dnsZoneETag/*IfNoneMatch*/);
-                    } else {
-                        return self.manager().inner().zones().createOrUpdateAsync(self.resourceGroupName(),
-                                self.name(), self.inner(), self.dnsZoneETag/*IfMatch*/, null/*IfNoneMatch*/);
-                    }
-                })
+                .flatMap(self -> self.manager().inner().zones().createOrUpdateAsync(self.resourceGroupName(),
+                        self.name(), self.inner(), this.eTagState.ifMatchValueOnUpdate(self.inner().etag()), eTagState.ifNonMatchValueOnCreate()))
                 .map(innerToFluentMap(this))
                 .map(dnsZone -> {
-                    this.dnsZoneETag = null;
+                    this.eTagState.clear();
                     return dnsZone;
                 });
     }
